Add regex and file support for http mocking, break out ContactHistory
