+++ temba-modules.ts
@@ -11,7 +11,8 @@ import { Loading } from './src/loading/Loading';
 import { CharCount } from './src/charcount/CharCount';
 import { Options } from './src/options/Options';
 import { ContactChat } from './src/contacts/ContactChat';
-import { TicketList } from './src/list/TicketList';
+import { ContactHistory } from './src/contacts/ContactHistory';
+import { ContactList } from './src/list/ContactList';
 import { ContactDetails } from './src/contacts/ContactDetails';
 import { TembaList } from './src/list/TembaList';
 import { ContactSearch } from './src/contactsearch/ContactSearch';
@@ -34,9 +35,10 @@ window.customElements.define('temba-field', FormField);
 window.customElements.define('temba-dialog', Dialog);
 window.customElements.define('temba-modax', Modax);
 window.customElements.define('temba-charcount', CharCount);
+window.customElements.define('temba-contact-history', ContactHistory);
 window.customElements.define('temba-contact-chat', ContactChat);
 window.customElements.define('temba-contact-details', ContactDetails);
-window.customElements.define('temba-tickets', TicketList);
+window.customElements.define('temba-contacts', ContactList);
 window.customElements.define('temba-list', TembaList);
 window.customElements.define('temba-contact-search', ContactSearch);
 window.customElements.define('temba-icon', VectorIcon);
