+++ test/temba-contact-chat.test.ts
@@ -1,17 +1,16 @@
 import { fixture, assert } from '@open-wc/testing';
 import { ContactChat } from '../src/contacts/ContactChat';
+import { getHTML } from '../test/utils.test';
 import './utils.test';
 
-export const getHTML = () => {
-  return `<temba-contact-chat></temba-contact-chat>`;
-};
+const getChatHTML = (attrs: any = {}) => getHTML('temba-contact-chat', attrs);
 
 describe('temba-contact-chat', () => {
   beforeEach(() => {});
   afterEach(() => {});
 
   it('can be created', async () => {
-    const chat: ContactChat = await fixture(getHTML());
+    const chat: ContactChat = await fixture(getChatHTML());
     assert.instanceOf(chat, ContactChat);
   });
 });
